Move validators to own folder (close #59)
